backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
google-genai
python-dotenv
numpy
requests
beautifulsoup4